Move PackageSettings, CSVSettings to /settings.
