+++ lib/arb_generator.dart
@@ -1,5 +1,5 @@
 library arb_generator;
 
-export 'src/models/csv_settings.dart';
-export 'src/models/package_settings.dart';
+export 'src/models/settings/csv_settings.dart';
+export 'src/models/settings/package_settings.dart';
 export 'src/services/arb_generator.dart';
